Revert "use termion's built-in `terminal_size` function"
This reverts commit 3f4188faa328f5770ba62870a850cdde3b55b048.

Unfortunately Termion's `terminal_size()` function requests the size of
whatever stdout is, so if you pipe Star's output to another program, or
try to write it to a file, you'll get an "Inappropriate ioctl for
device" error. The solution is to go back to our old `terminal_size()`
implementation, which is shamelessly ripped off of Termion's, but takes
a FD parameter. We can then provide the console's FD, which is not
necessarily stdout, to avoid this error.