Revert "use termion's built-in `terminal_size` function"
This reverts commit 0e89296bf4e5b6780bb2fba6e70f44c3faf7ad8b.

Unfortunately Termion's `terminal_size()` function requests the size of
whatever stdout is, so if you pipe Star's output to another program, or
try to write it to a file, you'll get an "Inappropriate ioctl for
device" error. The solution is to go back to our old `terminal_size()`
implementation, which is shamelessly ripped off of Termion's, but takes
a FD parameter. We can then provide the console's FD, which is not
necessarily stdout, to avoid this error.